+++ CHANGELOG.md
@@ -41,3 +41,4 @@ This release has a big theming change as well as including the move back to rofi
 - Hyprland plugins are now disabled in the config by default.
 - Fastfetch is now replacing neofetch.
 - Btop is back baby!
+- Switching to Brave as the default to protect user privacy.